requirements.txt
Django==4.1.7
pushy_python==1.0.9
